requirements.txt
Flask==2.3.2
mysql-connector-python==8.0.33
Flask-Login==0.6.2
Werkzeug==2.3.6
